Fixed error: creation of section and blocks folders
