+++ src/components/course/blocks/blockXmlFiles/generateInforefXml.js
@@ -6,6 +6,12 @@ const xmlbuilder = require('xmlbuilder');
 // Generates inforef.xml inside 'blocks' directory inside 'completion_progress'
 // course\blocks\completion_progress\inforef.xml
 function  generateInforefXml(completionProgressDir){
+    
+    // Ensure the directory exists
+    if (!fs.existsSync(completionProgressDir)) {
+        fs.mkdirSync(completionProgressDir, { recursive: true });
+    }
+
     const inforefXml = xmlbuilder.create('inforef', { encoding: 'UTF-8' })
     .end({ pretty: true });
     fs.writeFileSync(path.join(completionProgressDir, 'inforef.xml'), inforefXml);
