+++ src/components/course/blocks/generateBlocksFiles.js
@@ -1,14 +1,16 @@
-const { generateBlockXml } = require("./xmlFiles/generateBlockXml");
-const { generateInforefXml } = require("./xmlFiles/generateInforefXml");
-const { generateRolesXml } = require("./xmlFiles/generateRolesXml");
-
 // Generates xml files inside 'course' directory inside 'blocks' folder e.g completion_progress
+const { generateBlockXml } = require("./blockXmlFiles/generateBlockXml");
+const { generateInforefXml } = require("./blockXmlFiles/generateInforefXml");
+const { generateRolesXml } = require("./blockXmlFiles/generateRolesXml");
+
 // course\blocks\completion_progress
-function generateBlockFiles(completionProgressDir){
-    generateBlockXml(completionProgressDir)
-    generateInforefXml(completionProgressDir)
-    generateRolesXml(completionProgressDir)
+function generateBlockFiles (completionProgressDir) {
+    generateBlockXml(completionProgressDir);
+    generateRolesXml(completionProgressDir);
+    generateInforefXml(completionProgressDir);
+
 }
 
-module.exports = { generateBlockFiles }
+module.exports = { generateBlockFiles };
+
