+++ src/components/course/courseXmlFiles/generateCalendarXml.js
