+++ src/components/course/courseXmlFiles/generateCompetenciesXml.js
