+++ src/components/course/courseXmlFiles/generateCompletiondefaultsXml.js
