+++ src/components/course/courseXmlFiles/generateContentbankXml.js
