+++ src/components/course/courseXmlFiles/generateCourseXml.js
