+++ src/components/course/courseXmlFiles/generateEnrolmentXml.js
