+++ src/components/course/courseXmlFiles/generateFiltersXml.js
