+++ src/components/course/courseXmlFiles/generateInforefXml.js
