+++ src/components/course/courseXmlFiles/generateRolesXml.js
