+++ src/components/course/generateCourseFiles.js
@@ -1,12 +1,12 @@
-const { generateCalendarXml } = require("./xmlFiles/generateCalendarXml");
-const { generateCompetenciesXml } = require("./xmlFiles/generateCompetenciesXml");
-const { generateCompletiondefaultsXml } = require("./xmlFiles/generateCompletiondefaultsXml");
-const { generateContentbankXml } = require("./xmlFiles/generateContentbankXml");
-const { generateCourseXml } = require("./xmlFiles/generateCourseXml");
-const { generateEnrolmentXml } = require("./xmlFiles/generateEnrolmentXml");
-const { generateInforefXml } = require("./xmlFiles/generateInforefXml");
-const { generateFiltersXml } = require("./xmlFiles/generateFiltersXml");
-const { generateRolesXml } = require("./xmlFiles/generateRolesXml");
+const { generateCalendarXml } = require("./courseXmlFiles/generateCalendarXml");
+const { generateCompetenciesXml } = require("./courseXmlFiles/generateCompetenciesXml");
+const { generateCompletiondefaultsXml } = require("./courseXmlFiles/generateCompletiondefaultsXml");
+const { generateContentbankXml } = require("./courseXmlFiles/generateContentbankXml");
+const { generateCourseXml } = require("./courseXmlFiles/generateCourseXml");
+const { generateEnrolmentXml } = require("./courseXmlFiles/generateEnrolmentXml");
+const { generateInforefXml } = require("./courseXmlFiles/generateInforefXml");
+const { generateFiltersXml } = require("./courseXmlFiles/generateFiltersXml");
+const { generateRolesXml } = require("./courseXmlFiles/generateRolesXml");
 
 // Generates xml files inside 'course' directory
 function generateCourseFiles (courseDir) {
